+++ test/Core/State/Model/StateSpec.hs
@@ -1,10 +1,9 @@
 module Core.State.Model.StateSpec where
 
-import Core.State.Model.State 
-import Test.Hspec
+import           Core.State.Model.State
+import           Test.Hspec
 
 spec :: Spec
-spec =
-  describe "initialState" $ 
-    it "should be empty" $
-      initialState `shouldBe` State Nothing Nothing Nothing Nothing
+spec = describe "initialState" $ do
+  it "should be empty" $ do
+    initialState `shouldBe` State Nothing Nothing Nothing Nothing MonitorOff