+++ ChangeLog.md
@@ -3,3 +3,5 @@
 - [0.1.0] first fixed version
 
 ## Unreleased changes
+
+- Finite State Machine House Monitoring